Rename location.
